Infinispan transformers fixes following the rebase onto the improved tests
