+++ clustering/infinispan/src/test/java/org/jboss/as/clustering/infinispan/subsystem/InfinispanSubsystemTransformerTestCase.java
@@ -142,17 +142,10 @@ public class InfinispanSubsystemTransformerTestCase extends OperationTestCaseBas
 
         // cache attributes
         for (int i=0; i < cachePaths.length; i++) {
-//            config.addFailedAttribute(subsystemAddress.append(CacheContainerResource.CONTAINER_PATH).append(cachePaths[i]),
-//                    new FailedOperationTransformationConfig.RejectExpressionsConfig(
-//                            InfinispanRejectedExpressions_1_3.ACCEPT14_REJECT13_CACHE_ATTRIBUTES).setNotExpectedWriteFailure(ModelKeys.SEGMENTS));
-
             config.addFailedAttribute(subsystemAddress.append(CacheContainerResource.CONTAINER_PATH).append(cachePaths[i]),
-//                    new FailedOperationTransformationConfig.RejectExpressionsConfig(
-//                            // InfinispanRejectedExpressions_1_3.ACCEPT14_REJECT13_CACHE_ATTRIBUTES).setNotExpectedWriteFailure(ModelKeys.INDEXING_PROPERTIES, ModelKeys.SEGMENTS));
-//                            InfinispanRejectedExpressions_1_3.ACCEPT14_REJECT13_CACHE_ATTRIBUTES).setNotExpectedWriteFailure(ModelKeys.INDEXING_PROPERTIES));
                     FailedOperationTransformationConfig.ChainedConfig.createBuilder(InfinispanRejectedExpressions_1_3.ACCEPT14_REJECT13_CACHE_ATTRIBUTES)
                     .addConfig(new FailedOperationTransformationConfig.RejectExpressionsConfig(InfinispanRejectedExpressions_1_3.ACCEPT14_REJECT13_CACHE_ATTRIBUTES))
-                    .addConfig(new RemoveResolvedIndexingPropertiesConfig(CacheResource.INDEXING_PROPERTIES)).build().setNotExpectedWriteFailure(ModelKeys.SEGMENTS));
+                    .addConfig(new RemoveResolvedIndexingPropertiesConfig(CacheResource.INDEXING_PROPERTIES)).build()/*.setNotExpectedWriteFailure(ModelKeys.SEGMENTS)*/);
 
             PathElement[] childPaths = {
                     LockingResource.LOCKING_PATH,
@@ -215,8 +208,7 @@ public class InfinispanSubsystemTransformerTestCase extends OperationTestCaseBas
         @Override
         protected boolean checkValue(String attrName, ModelNode attribute, boolean isWriteAttribute) {
             //The add does not currently reject the defined indexing-properties
-            //TODO once this is moved to builders, the add should also remove the defined indexing-properties
-            if (isWriteAttribute && attribute.isDefined() && attrName.equals(CacheResource.INDEXING_PROPERTIES.getName())) {
+            if (attribute.isDefined() && attrName.equals(CacheResource.INDEXING_PROPERTIES.getName())) {
                 ModelNode resolved = attribute.resolve();
                 return resolved.equals(attribute);
             }
