+++ model-test/src/main/java/org/jboss/as/model/test/FailedOperationTransformationConfig.java
@@ -203,6 +203,7 @@ public class FailedOperationTransformationConfig {
         /**
          * Whether it is expected that the following operation should fail
          *
+         * @param operation the operation to check
          * @return {@code true} if expected to fail
          */
         boolean expectFailed(ModelNode operation);
@@ -211,6 +212,7 @@ public class FailedOperationTransformationConfig {
          * Whether something can be corrected in the operation to make it pass.
          * It is preferable to correct one attribute at a time.
          *
+         * @param the operation to check
          * @return {@code true} if expected to fail, {@code false} otherwise
          */
         boolean canCorrectMore(ModelNode operation);
@@ -219,6 +221,7 @@ public class FailedOperationTransformationConfig {
          * Correct the operation, only called if {@link #canCorrectMore(ModelNode)} returned {@code true}
          * It is preferable to correct one attribute at a time.
          *
+         * @param operation the operation to correct
          * @return the corrected operation
          */
         ModelNode correctOperation(ModelNode operation);
@@ -231,13 +234,17 @@ public class FailedOperationTransformationConfig {
         /**
          * Whether it is expected that the following write attribute operation should fail
          *
+         * @param the 'add' operation to correct
          * @return {@code true} if expected to fail
          */
         boolean expectFailedWriteAttributeOperation(ModelNode operation);
 
         /**
-         * Correct the operation.
+         * Correct the operation. This only deals with one attribute, and
+         * the framework will only call this once if it failed and {@link #correctWriteAttributeOperation(ModelNode)}
+         * returned {@code true}, so make sure to do everything to correct the {@code value} attribute.
          *
+         * @param the 'write-attribute' operation to correct
          * @return the corrected operation
          */
         ModelNode correctWriteAttributeOperation(ModelNode operation);
@@ -245,6 +252,7 @@ public class FailedOperationTransformationConfig {
         /**
          * Whether the operation is expected to be discarded
          *
+         * @param operation the operation to check
          * @return {@code true} if expected to fail
          */
         boolean expectDiscarded(ModelNode operation);
